Update the ember configure the locationType to has instead of auto. When in ‘auto’, the URL path would be something like ‘/user/login’. When in ‘hash’, the URL path would be something like ‘/user/#/login’.
In first ‘auto’ model, server container will treat it as sub folder hence it would not have folder /login under /user so we will get 404 when doing refresh from page.
When doing hash, server container will look at ‘/user’, then use ‘#/login’ as mark in the page. In /user, we have index.html which is severed as default page. So we can find the page.